+++ src/components/BadgerBadge/BadgerBadge.js
@@ -23,6 +23,7 @@ import SLPLogoImage from '../../images/slp-logo.png';
 import colors from '../../styles/colors';
 
 import PriceDisplay from '../PriceDisplay';
+import InvoiceTimer from '../InvoiceTimer';
 
 import Button from '../../atoms/Button';
 import ButtonQR from '../../atoms/ButtonQR';
@@ -109,6 +110,10 @@ type Props = BadgerBaseProps & {
 	showQR?: boolean,
 	showBorder?: boolean,
 
+	invoiceInfo: ?Object,
+	invoiceTimeLeftSeconds: ?number,
+	invoiceFiat: ?number,
+
 	handleClick: Function,
 };
 
@@ -144,33 +149,77 @@ class BadgerBadge extends React.PureComponent<Props> {
 			showAmount,
 			showQR,
 			paymentRequestUrl,
+			invoiceInfo,
+			invoiceTimeLeftSeconds,
+			invoiceFiat,
 			showBorder,
 			showBrand,
 		} = this.props;
 
 		const CoinImage = coinType === 'BCH' ? BitcoinCashImage : SLPLogoImage;
+
+		// Handle cases for displaying price
+		// Case 1: no bip70 invoice
+		// Case 2: bip70 invoice, supported pay.bitcoin.com invoice with websocket info over-riding props, showAmount={true}
+		// Case 3: bip70 invoice, no supported websocket price updating, only display a "Bip70 invoice" label and no cointype or price, OR valid invoice but showAmount={false}; a bip70 invoice must be labeled as such even if showAmount={false} and price in props must be over-ridden
+
+		// Case 1: no bip70 invoice
+		let displayedPriceInfo = (
+			<Prices>
+				{price != undefined && (
+					<PriceDisplay
+						preSymbol={getCurrencyPreSymbol(currency)}
+						price={formatPriceDisplay(price)}
+						symbol={currency}
+					/>
+				)}
+				{showAmount && (
+					<PriceDisplay
+						coinType={coinType}
+						price={formatAmount(amount, coinDecimals)}
+						symbol={coinSymbol}
+						name={coinName}
+					/>
+				)}
+			</Prices>
+		);
+		// Case 2: bip70 invoice, supported pay.bitcoin.com invoice with websocket info over-riding props, showAmount={true}
+		if (showAmount && paymentRequestUrl && invoiceInfo.currency) {
+			displayedPriceInfo = (
+				<Prices>
+					{invoiceFiat != undefined && (
+						<PriceDisplay
+							preSymbol={getCurrencyPreSymbol(currency)}
+							price={formatPriceDisplay(invoiceFiat)}
+							symbol={currency}
+						/>
+					)}
+
+					<PriceDisplay
+						coinType={coinType}
+						price={formatAmount(amount, coinDecimals)}
+						symbol={coinSymbol}
+						name={coinName}
+					/>
+				</Prices>
+			);
+			// Case 3: bip70 invoice, no supported websocket price updating, only display a "Bip70 invoice" label and no cointype or price, OR valid invoice but showAmount={false}
+			// Bip70 invoice must be labeled as such if amounts are not shown
+		} else if (paymentRequestUrl) {
+			displayedPriceInfo = (
+				<Prices>
+					<PriceText>BIP70 Invoice</PriceText>
+				</Prices>
+			);
+		}
 		return (
 			<Outer>
 				<Main showBorder={showBorder}>
 					<H3>{text}</H3>
-					<Prices>
-						{price != undefined && (
-							<PriceDisplay
-								preSymbol={getCurrencyPreSymbol(currency)}
-								price={formatPriceDisplay(price)}
-								symbol={currency}
-							/>
-						)}
-						{showAmount && (
-							<PriceDisplay
-								coinType={coinType}
-								price={formatAmount(amount, coinDecimals)}
-								symbol={coinSymbol}
-								name={coinName}
-								paymentRequestUrl={paymentRequestUrl}
-							/>
-						)}
-					</Prices>
+					{displayedPriceInfo}
+					{invoiceTimeLeftSeconds !== null && (
+						<InvoiceTimer invoiceTimeLeftSeconds={invoiceTimeLeftSeconds} />
+					)}
 					<ButtonContainer>
 						{showQR ? (
 							<ButtonQR