+++ src/hoc/BadgerBase/BadgerBase.js
@@ -6,10 +6,19 @@ import debounce from 'lodash/debounce';
 
 import {
 	fiatToSatoshis,
+	bchToFiat,
 	adjustAmount,
 	getAddressUnconfirmed,
 	getTokenInfo,
 } from '../../utils/badger-helpers';
+
+import {
+	getWalletProviderStatus,
+	constants,
+	sendAssets,
+	payInvoice,
+} from 'bitcoin-wallet-api';
+const { WalletProviderStatus } = constants;
 
 import { type CurrencyCode } from '../../utils/currency-helpers';
 
@@ -24,7 +33,13 @@ const URI_CHECK_INTERVAL = 10 * SECOND;
 type ValidCoinTypes = 'BCH' | 'SLP';
 
 // TODO - Login/Install are badger states, others are payment states.  Separate them to be independent
-type ButtonStates = 'fresh' | 'pending' | 'complete' | 'login' | 'install';
+type ButtonStates =
+	| 'fresh'
+	| 'pending'
+	| 'complete'
+	| 'expired'
+	| 'login'
+	| 'install';
 
 type BadgerBaseProps = {
 	to: string,
@@ -58,15 +73,22 @@ type State = {
 	errors: string[],
 
 	satoshis: ?number, // Used when converting fiat to BCH
+	invoiceFiat: ?number, // Used to show USD cost of a BCH BIP70 invoice
 
 	coinSymbol: ?string,
 	coinName: ?string,
 	coinDecimals: ?number,
 	unconfirmedCount: ?number,
+	invoiceInfo: ?Object,
+	invoiceTimeLeftSeconds: ?number,
 
 	intervalPrice: ?IntervalID,
+	intervalInvoicePrice: ?IntervalID,
 	intervalLogin: ?IntervalID,
 	intervalUnconfirmed: ?IntervalID,
+	intervalTimer: ?IntervalId,
+
+	websocketInvoice: ?Object,
 };
 
 const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
@@ -90,11 +112,18 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 			coinName: null,
 
 			unconfirmedCount: null,
+			invoiceInfo: {},
+			invoiceFiat: null,
+			invoiceTimeLeftSeconds: null,
 
 			intervalPrice: null,
+			intervalInvoicePrice: null,
 			intervalLogin: null,
 			intervalUnconfirmed: null,
+			intervalTimer: null,
 			errors: [],
+
+			websocketInvoice: null,
 		};
 
 		addError = (error: string) => {
@@ -109,7 +138,11 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 
 		paymentSendSuccess = () => {
 			const { isRepeatable } = this.props;
-			const { intervalUnconfirmed, unconfirmedCount } = this.state;
+			const {
+				intervalUnconfirmed,
+				unconfirmedCount,
+				intervalTimer,
+			} = this.state;
 
 			this.setState({
 				step: 'complete',
@@ -121,6 +154,18 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 			} else {
 				intervalUnconfirmed && clearInterval(intervalUnconfirmed);
 			}
+			// If invoice is paid, clear timer, and set secondsLeft to null to hide clock
+			intervalTimer && clearInterval(intervalTimer);
+			this.setState({ invoiceTimeLeftSeconds: null });
+		};
+
+		invoiceExpired = () => {
+			const { intervalTimer } = this.state;
+			this.setState({
+				step: 'expired',
+			});
+			intervalTimer && clearInterval(intervalTimer);
+			this.setState({ invoiceTimeLeftSeconds: null });
 		};
 
 		handleClick = () => {
@@ -143,74 +188,70 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 				return;
 			}
 
+			const walletProviderStatus = getWalletProviderStatus();
+
 			if (
-				typeof window !== `undefined` &&
-				typeof window.Web4Bch !== 'undefined'
+				typeof window === `undefined` ||
+				(walletProviderStatus.badger === WalletProviderStatus.NOT_AVAILABLE &&
+					walletProviderStatus.android === WalletProviderStatus.NOT_AVAILABLE &&
+					walletProviderStatus.ios === WalletProviderStatus.NOT_AVAILABLE)
 			) {
-				const { web4bch } = window;
-
-				const web4bch2 = new window.Web4Bch(web4bch.currentProvider);
-				const { defaultAccount } = web4bch2.bch;
-
-				if (!defaultAccount) {
-					this.setState({ step: 'login' });
-					return;
-				}
-
-				// BCH amount = satoshis, SLP amount = absolute value
-				const calculatedValue =
-					coinType === 'BCH' && amount
-						? adjustAmount(amount, 8)
-						: amount || satoshis;
-
-				const txParamsBase = {
-					to,
-					from: defaultAccount,
-					value: calculatedValue,
-				};
-
-				const txParamsSLP =
-					coinType === 'SLP' && tokenId
-						? {
-								...txParamsBase,
-								sendTokenData: {
-									tokenId,
-									tokenProtocol: 'slp',
-								},
-						  }
-						: txParamsBase;
-
-				const txParamsOpReturn =
-					opReturn && opReturn.length
-						? { ...txParamsSLP, opReturn: { data: opReturn } }
-						: txParamsSLP;
-
-				const txParams =
-					paymentRequestUrl && paymentRequestUrl.length
-						? { paymentRequestUrl } // If there is an invoice, this will be the only txParams
-						: txParamsOpReturn;
-
-				this.setState({ step: 'pending' });
-
-				console.info('Badger send begin', txParams);
-				web4bch2.bch.sendTransaction(txParams, (err, res) => {
-					if (err) {
-						console.info('Badger send cancel', err);
-						failFn && failFn(err);
-						this.setState({ step: 'fresh' });
-					} else {
-						console.info('Badger send success:', res);
-						successFn && successFn(res);
-						this.paymentSendSuccess();
-					}
-				});
-			} else {
 				this.setState({ step: 'install' });
 
 				if (typeof window !== 'undefined') {
 					window.open('https://badger.bitcoin.com');
 				}
+				return;
 			}
+
+			if (walletProviderStatus.badger === WalletProviderStatus.AVAILABLE) {
+				this.setState({ step: 'login' });
+				return;
+			}
+
+			if (paymentRequestUrl) {
+				this.setState({ step: 'pending' });
+				console.info('Badger payInvoice begin', paymentRequestUrl);
+				payInvoice({ url: paymentRequestUrl })
+				.then(({ memo }) => {
+					console.info('Badger send success:', memo);
+					successFn && successFn(memo);
+					this.paymentSendSuccess();
+				})
+				.catch((err) => {
+					console.info('Badger send cancel', err);
+					failFn && failFn(err);
+					this.setState({ step: 'fresh' });
+				});
+				return;
+			}
+
+			const sendParams = {
+				to,
+				protocol: coinType,
+				value: amount || adjustAmount(satoshis, 8, true),
+			};
+			if (coinType === 'SLP') {
+				sendParams.assetId = tokenId;
+			}
+
+			if (opReturn && opReturn.length) {
+				sendParams.opReturn = opReturn;
+			}
+			
+			this.setState({ step: 'pending' });
+			console.info('Badger sendAssets begin', sendParams);
+			sendAssets(sendParams)
+			.then(({txid}) => {
+				console.info('Badger send success:', txid);
+				successFn && successFn(txid);
+				this.paymentSendSuccess();
+			})
+			.catch(err => {
+				console.info('Badger send cancel', err);
+				failFn && failFn(err);
+				this.setState({ step: 'fresh' });
+			});
 		};
 
 		gotoLoginState = () => {
@@ -218,10 +259,12 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 			this.setState({ step: 'login' });
 			if (typeof window !== 'undefined') {
 				const intervalLogin = setInterval(() => {
-					const { web4bch } = window;
-					const web4bch2 = new window.Web4Bch(web4bch.currentProvider);
-					const { defaultAccount } = web4bch2.bch;
-					if (defaultAccount) {
+					const walletProviderStatus = getWalletProviderStatus();
+					if (
+						walletProviderStatus.badger === WalletProviderStatus.LOGGED_IN ||
+						walletProviderStatus.android === WalletProviderStatus.AVAILABLE ||
+						walletProviderStatus.ios === WalletProviderStatus.AVAILABLE
+					) {
 						clearInterval(intervalLogin);
 						this.setState({ step: 'fresh' });
 					}
@@ -256,6 +299,33 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 			this.setState({ intervalPrice: intervalPriceNext });
 		};
 
+		updateInvoiceFiat = debounce(
+			async () => {
+				const { currency } = this.props;
+				const { invoiceInfo } = this.state;
+				const invoicePriceBCH = invoiceInfo.fiatTotal;
+
+				if (!invoiceInfo.fiatTotal || invoiceInfo.currency !== 'BCH') return;
+				const invoiceFiat = await bchToFiat(currency, invoicePriceBCH);
+				this.setState({ invoiceFiat });
+			},
+			250,
+			{ lead: true, trailing: true }
+		);
+
+		setupInvoiceFiat = () => {
+			const { intervalInvoicePrice } = this.state;
+			intervalInvoicePrice && clearInterval(intervalInvoicePrice);
+
+			this.updateInvoiceFiat();
+			const intervalInvoicePriceNext = setInterval(
+				() => this.updateInvoiceFiat(),
+				PRICE_UPDATE_INTERVAL
+			);
+
+			this.setState({ intervalInvoicePrice: intervalInvoicePriceNext });
+		};
+
 		setupWatchAddress = async () => {
 			const { to } = this.props;
 			const { intervalUnconfirmed } = this.state;
@@ -284,15 +354,96 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 			this.setState({ intervalUnconfirmed: intervalUnconfirmedNext });
 		};
 
-		setupCoinMeta = async () => {
-			const { coinType, tokenId } = this.props;
-			if (coinType === 'BCH') {
+		setupWatchInvoice = async () => {
+			const { paymentRequestUrl, successFn } = this.props;
+
+			const urlParts = paymentRequestUrl.split('/');
+			const server = urlParts[2];
+			if (server !== 'pay.bitcoin.com') {
+				// InvoiceTimer and showAmount fields are only supported for pay.bitcoin.com invoices
+				return;
+			}
+
+			const paymentId = urlParts[urlParts.length - 1];
+			const ws = new WebSocket(`wss://pay.bitcoin.com/s/${paymentId}`);
+
+			ws.onopen = () => {
+				this.setState({ websocketInvoice: ws });
+			};
+
+			ws.onmessage = (evt) => {
+				// listen to data sent from the websocket server
+				const invoiceInfo = JSON.parse(evt.data);
+
+				const invoiceStatus = invoiceInfo.status; // for InvoiceDisplay
+
+				this.setState({ invoiceInfo }, this.setupInvoiceFiat());
+				this.setupCoinMeta(invoiceInfo);
+
+				if (invoiceStatus === 'paid') {
+					successFn && successFn();
+					return this.paymentSendSuccess();
+				}
+				if (invoiceStatus === 'expired') {
+					return this.invoiceExpired();
+				}
+
+				// If invoice is not expired or paid, start the timer (add this logic after timer works) TODO
+				// Get current UTC time for timer
+				const nowUTC = Date.now();
+
+				const invoiceExpiresAt = Date.parse(invoiceInfo.expires);
+
+				let invoiceTimeLeftSeconds = Math.round(
+					(invoiceExpiresAt - nowUTC) / 1000
+				);
+
+				if (invoiceTimeLeftSeconds > 0) {
+					this.setState({ invoiceTimeLeftSeconds });
+				}
+			};
+		};
+
+		setupInvoiceTimer = () => {
+			// start timer
+			const intervalTimerNext = setInterval(() => {
+				const prevInvoiceTimeLeftSeconds = this.state.invoiceTimeLeftSeconds;
+				if (prevInvoiceTimeLeftSeconds === 1) {
+					return this.invoiceExpired();
+				}
+				const newInvoiceTimeLeftSeconds = prevInvoiceTimeLeftSeconds - 1;
+				this.setState({
+					invoiceTimeLeftSeconds: newInvoiceTimeLeftSeconds,
+				});
+			}, 1000);
+
+			this.setState({ intervalTimer: intervalTimerNext });
+		};
+
+		setupCoinMeta = async (invoiceInfo = null) => {
+			const { coinType, tokenId, paymentRequestUrl } = this.props;
+
+			if (invoiceInfo !== null && invoiceInfo.currency !== 'BCH') {
+				const invoiceTokenId = invoiceInfo.outputs[0].token_id;
+				const invoiceTokenInfo = await getTokenInfo(invoiceTokenId);
+
+				const { symbol, decimals, name } = invoiceTokenInfo;
+
+				this.setState({
+					coinSymbol: symbol,
+					coinDecimals: decimals,
+					coinName: name,
+				});
+			} else if (
+				(!paymentRequestUrl && coinType === 'BCH') ||
+				(invoiceInfo !== null && invoiceInfo.currency === 'BCH')
+			) {
 				this.setState({
 					coinSymbol: 'BCH',
 					coinDecimals: 8,
 					coinName: 'Bitcoin Cash',
 				});
-			} else if (coinType === 'SLP' && tokenId) {
+			} else if (!paymentRequestUrl && coinType === 'SLP' && tokenId) {
 				this.setState({
 					coinSymbol: null,
 					coinName: null,
@@ -311,35 +462,53 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 
 		async componentDidMount() {
 			if (typeof window !== 'undefined') {
-				const { price, coinType, amount, watchAddress } = this.props;
+				const {
+					price,
+					coinType,
+					amount,
+					watchAddress,
+					paymentRequestUrl,
+				} = this.props;
 
 				// setup state, intervals, and listeners
 				watchAddress && this.setupWatchAddress();
+				paymentRequestUrl && this.setupWatchInvoice(); // sets up websocket for invoice which calls setupCoinMeta() when the necessary information has been received
 				price && this.setupSatoshisFiat();
-				this.setupCoinMeta();
+				!paymentRequestUrl && this.setupCoinMeta(); // normal call for setupCoinMeta()
 
 				// Detect Badger and determine if button should show login or install CTA
-				if (window.Web4Bch) {
-					const { web4bch } = window;
-					const web4bch2 = new window.Web4Bch(web4bch.currentProvider);
-					const { defaultAccount } = web4bch2.bch;
-					if (!defaultAccount) {
-						this.gotoLoginState();
-					}
-				} else {
+				const walletProviderStatus = getWalletProviderStatus();
+				if (walletProviderStatus.badger === WalletProviderStatus.AVAILABLE) {
+					this.gotoLoginState();
+				}
+				if (
+					walletProviderStatus.badger === WalletProviderStatus.NOT_AVAILABLE &&
+					walletProviderStatus.android === WalletProviderStatus.NOT_AVAILABLE &&
+					walletProviderStatus.ios === WalletProviderStatus.NOT_AVAILABLE
+				) {
 					this.setState({ step: 'install' });
 				}
 			}
 		}
 
 		componentWillUnmount() {
-			const { intervalPrice, intervalLogin, intervalUnconfirmed } = this.state;
+			const {
+				intervalPrice,
+				intervalInvoicePrice,
+				intervalLogin,
+				intervalUnconfirmed,
+				websocketInvoice,
+				intervalTimer,
+			} = this.state;
 			intervalPrice && clearInterval(intervalPrice);
+			intervalInvoicePrice && clearInterval(intervalInvoicePrice);
 			intervalLogin && clearInterval(intervalLogin);
 			intervalUnconfirmed && clearInterval(intervalUnconfirmed);
+			intervalTimer && clearInterval(intervalTimer);
+			websocketInvoice && websocketInvoice.close();
 		}
 
-		componentDidUpdate(prevProps: BadgerBaseProps) {
+		componentDidUpdate(prevProps: BadgerBaseProps, prevState) {
 			if (typeof window !== 'undefined') {
 				const {
 					currency,
@@ -349,8 +518,14 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 					isRepeatable,
 					watchAddress,
 					tokenId,
+					paymentRequestUrl,
 				} = this.props;
-				const { intervalPrice } = this.state;
+				const {
+					invoiceTimeLeftSeconds,
+					websocketInvoice,
+					intervalInvoicePrice,
+					intervalTimer,
+				} = this.state;
 
 				const prevCurrency = prevProps.currency;
 				const prevCoinType = prevProps.coinType;
@@ -359,6 +534,30 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 				const prevIsRepeatable = prevProps.isRepeatable;
 				const prevWatchAddress = prevProps.watchAddress;
 				const prevTokenId = prevProps.tokenId;
+				const prevPaymentRequestUrl = prevProps.paymentRequestUrl;
+
+				const prevInvoiceTimeLeftSeconds = prevState.invoiceTimeLeftSeconds;
+
+				if (paymentRequestUrl !== prevPaymentRequestUrl) {
+					// clear all invoice intervals and websockets
+					intervalTimer && clearInterval(intervalTimer);
+					intervalInvoicePrice && clearInterval(intervalInvoicePrice);
+					websocketInvoice && websocketInvoice.close();
+					// reset all invoice info, then set it up again
+					this.setState({
+						step: 'fresh',
+						invoiceInfo: {},
+						invoiceFiat: null,
+						invoiceTimeLeftSeconds: null,
+					});
+					this.setupWatchInvoice();
+				}
+				if (
+					invoiceTimeLeftSeconds !== null &&
+					prevInvoiceTimeLeftSeconds === null
+				) {
+					this.setupInvoiceTimer();
+				}
 
 				// Fiat price or currency changes
 				if (currency !== prevCurrency || price !== prevPrice) {
@@ -369,7 +568,7 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 					this.startRepeatable();
 				}
 
-				if (tokenId !== prevTokenId) {
+				if (tokenId !== prevTokenId && !paymentRequestUrl) {
 					this.setupCoinMeta();
 				}
 
@@ -393,17 +592,49 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 				stepControlled,
 				paymentRequestUrl,
 			} = this.props;
-			const { step, satoshis, coinDecimals, coinSymbol, coinName } = this.state;
+			const {
+				step,
+				satoshis,
+				coinDecimals,
+				coinSymbol,
+				coinName,
+				invoiceInfo,
+				invoiceTimeLeftSeconds,
+				invoiceFiat,
+			} = this.state;
 
-			const calculatedAmount = adjustAmount(amount, coinDecimals) || satoshis;
+			let calculatedAmount = adjustAmount(amount, coinDecimals) || satoshis;
 
-			// Only show QR if all requested features can be encoded in the BIP44 URI
+			// If this is a BIP70 invoice with a specified BCH amount, use that
+			if (paymentRequestUrl && invoiceInfo.currency === 'BCH') {
+				calculatedAmount = adjustAmount(invoiceInfo.fiatTotal, coinDecimals);
+			} else if (paymentRequestUrl && invoiceInfo.currency === 'SLP') {
+				// Sum up the SLP amounts from invoiceInfo.outputs[0].send_amounts
+				const amounts = invoiceInfo.outputs[0].send_amounts;
+				calculatedAmount = 0;
+				for (let i = 0; i < amounts.length; i++) {
+					calculatedAmount += amounts[i];
+				}
+			}
+
+			// Only show QR if all requested features can be encoded in the BIP44 URI, or if it's a BIP70 invoice
 			const shouldShowQR =
-				showQR && coinType === 'BCH' && (!opReturn || !opReturn.length);
+				(showQR && coinType === 'BCH' && (!opReturn || !opReturn.length)) ||
+				paymentRequestUrl;
+
+			// Show SLP icon if BIP70 invoice is for SLP
+			let determinedCoinType = coinType;
+			if (paymentRequestUrl && coinSymbol !== 'BCH' && coinSymbol !== null) {
+				determinedCoinType = 'SLP';
+			}
+			if (paymentRequestUrl && coinSymbol === 'BCH') {
+				determinedCoinType = 'BCH';
+			}
 
 			return (
 				<Wrapped
 					{...this.props}
+					coinType={determinedCoinType}
 					showQR={shouldShowQR}
 					handleClick={this.handleClick}
 					step={stepControlled || step}
@@ -411,7 +642,9 @@ const BadgerBase = (Wrapped: React.AbstractComponent<any>) => {
 					coinDecimals={coinDecimals}
 					coinSymbol={coinSymbol}
 					coinName={coinName}
-					paymentRequestUrl={paymentRequestUrl}
+					invoiceInfo={invoiceInfo}
+					invoiceTimeLeftSeconds={invoiceTimeLeftSeconds}
+					invoiceFiat={invoiceFiat}
 				/>
 			);
 		}